Google Sign In completamente implementato nel sito. TODO: test: creazione nuovo utente da google e update utente esistente (+ googleId)

+++ app/public/js/status-auth.js
@@ -1,10 +1,24 @@
-// AUTH STATUS CHANGES LISTENER
-auth.onAuthStateChanged(user => {
-    if(user) {
-        console.log('user logged in: ',user);
-    } else {
-        console.log('user logged out');
-        localStorage.clear();          //LOCALSTORAGE REMOVE
-        window.location = '/login';
-    }
-});
+function gapiInit() {
+
+    gapi.load("auth2", function gapiInit() {
+
+        var googleOauthClient = gapi.auth2.getAuthInstance();
+
+        if (googleOauthClient == null) {
+
+            // AUTH STATUS CHANGES LISTENER
+            auth.onAuthStateChanged(user => {
+                if(user) {
+                    console.log('user logged in: ',user);
+                } else {
+                    console.log('user logged out');
+                    localStorage.clear();          //LOCALSTORAGE REMOVE
+                    window.location = '/login';
+                }
+            });
+
+        }
+
+    });
+    
+}
